+++ app/(main)/(screens)/channel/[cid].tsx
@@ -1,56 +1,13 @@
-import AddMember from "@/components/channel/add-member";
-import { View } from "@/components/theme/Themed";
-import ScreenHeader from "@/components/ui/screen-header";
-import { useLocalSearchParams } from "expo-router";
-import { useEffect, useState } from "react";
-import { ActivityIndicator } from "react-native";
-import { Channel as ChannelType } from "stream-chat";
-import { Channel, MessageInput, MessageList, useChatContext } from "stream-chat-expo";
+import ChannelListNative from "@/components/channel/channel-list";
+import ChannelListWeb from "@/components/channel/channel-list.web";
+import { Platform } from "react-native";
 
 const ChannelScreen = () => {
-  const [channel, setChannel] = useState<ChannelType | null>(null);
-  const { cid } = useLocalSearchParams<{ cid: string }>();
-  const [modalVisible, setModalVisible] = useState(false);
-
-  const { client } = useChatContext();
-
-  useEffect(() => {
-    const fetchChannel = async () => {
-      const channels = await client.queryChannels({ cid });
-      setChannel(channels[0]);
-    };
-
-    fetchChannel();
-  }, [cid]);
-
-  if (!channel) {
-    return (
-      <View
-        style={{
-          alignItems: 'center',
-          flex: 1,
-          justifyContent: 'center',
-        }}
-      >
-        <ActivityIndicator />
-      </View>
-    );
+  if (Platform.OS === 'web') {
+    return <ChannelListWeb />
   }
-
   return (
-    <Channel channel={channel} audioRecordingEnabled>
-      <ScreenHeader
-        title={channel?.data?.name || 'Chat'}
-        rightAction={() => setModalVisible(true)}
-      />
-      <MessageList />
-      <MessageInput />
-      <AddMember
-        channel={channel}
-        setVisible={setModalVisible}
-        visible={modalVisible}
-      />
-    </Channel>
+    <ChannelListNative />
   );
 }
