+++ components/BottomSheetActions.tsx
@@ -8,11 +8,11 @@ import {
 } from "react-native";
 import { List } from "react-native-paper";
 import BottomSheet, { BottomSheetView } from "@gorhom/bottom-sheet";
-import { useChatContext as useChat, useNotificationContext } from "@/contexts";
+import { useNotificationContext } from "@/contexts";
 import { FirestoreDB } from "@/utils/firestore";
 import { doc, updateDoc } from "firebase/firestore";
 import Toaster from "@/shared-uis/components/toaster/Toaster";
-import { useChatContext } from "stream-chat-expo";
+import { useChatHook } from "@/hooks/use-chat";
 
 interface BottomSheetActionsProps {
   cardType:
@@ -41,10 +41,10 @@ const BottomSheetActions = ({
   const [isMessageModalVisible, setIsMessageModalVisible] =
     React.useState(false);
 
-  const { client } = useChatContext();
   const {
     createGroupWithMembers,
-  } = useChat();
+    client,
+  } = useChatHook();
   const {
     createNotification,
   } = useNotificationContext();
@@ -74,7 +74,9 @@ const BottomSheetActions = ({
       await updateDoc(applicationRef, {
         status: "accepted",
       }).then(() => {
+        // @ts-ignore
         createGroupWithMembers(client, data.collaboration.name, [
+          // @ts-ignore
           client.user?.id as string,
           cardId.influencerID,
         ]).then(() => {
