+++ components/channel/channel.web.tsx
@@ -1,0 +1,17 @@
+import { Text, View } from "@/components/theme/Themed";
+
+const ChannelWeb = () => {
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center'
+      }}
+    >
+      <Text>Stream chat is not available on web right now</Text>
+    </View>
+  );
+};
+
+export default ChannelWeb;
