+++ constants/Placeholder.ts
@@ -1,4 +1,4 @@
 export const PLACEHOLDER_PERSON_IMAGE =
-  "/assets/images/placeholder-person-image.png";
+  "@/assets/images/placeholder-person-image.png";
 
-export const PLACEHOLDER_IMAGE = "/assets/images/placeholder-image.jpg";
+export const PLACEHOLDER_IMAGE = "@/assets/images/placeholder-image.jpg";
