+++ contexts/chat.provider.web.tsx
@@ -1,0 +1,11 @@
+import { PropsWithChildren } from "react";
+
+export const ChatProvider: React.FC<PropsWithChildren> = ({
+  children,
+}) => {
+  return (
+    <>
+      {children}
+    </>
+  );
+};
