+++ firebase-messaging-sw.js
@@ -1,0 +1,65 @@
+importScripts("https://www.gstatic.com/firebasejs/8.10.1/firebase-app.js");
+importScripts(
+  "https://www.gstatic.com/firebasejs/8.10.1/firebase-messaging.js"
+);
+
+const firebaseConfig = {
+  apiKey: "dummy",
+  authDomain: "dummy",
+  projectId: "dummy",
+  storageBucket: "dummy",
+  messagingSenderId: "dummy",
+  appId: "dummy",
+  measurementId: "dummy",
+};
+
+firebase.initializeApp(firebaseConfig);
+
+const messaging = firebase.messaging();
+
+messaging.onBackgroundMessage((payload) => {
+  console.log(
+    "[firebase-messaging-sw.js] Received background message ",
+    payload
+  );
+
+  const link = payload.fcmOptions?.link || payload.data?.link;
+
+  const notificationTitle = payload.notification?.title;
+  const notificationOptions = {
+    body: payload.notification?.body,
+    icon: payload.notification?.icon,
+    data: {
+      url: link,
+    },
+  };
+
+  self.registration.showNotification(notificationTitle, notificationOptions);
+});
+
+self.addEventListener("notificationclick", function (event) {
+  event.notification.close();
+
+  event.waitUntil(
+    clients
+      .matchAll({
+        type: "window",
+        includeUncontrolled: true,
+      })
+      .then(function (clientList) {
+        const url = event.notification.data.url;
+
+        if (!url) return;
+
+        for (const client of clientList) {
+          if (client.url === url && "focus" in client) {
+            return client.focus();
+          }
+        }
+
+        if (clients.openWindow) {
+          return clients.openWindow(url);
+        }
+      })
+  );
+});
