+++ hooks/use-chat.web.tsx
@@ -1,0 +1,15 @@
+import {
+  useChatContext as useChat,
+} from "@/contexts";
+
+export const useChatHookWeb = () => {
+  const client = '';
+  const {
+    createGroupWithMembers,
+  } = useChat();
+
+  return {
+    client,
+    createGroupWithMembers,
+  }
+}
