+++ package.web.json
@@ -65,8 +65,7 @@
     "react-native-swiper": "^1.6.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-video": "^6.7.0",
-    "react-native-web": "~0.19.10",
-    "stream-chat-expo": "^5.41.2"
+    "react-native-web": "~0.19.10"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
